+++ utils/Tile.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Tile.cpp                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: matteo <matteo@student.42.fr>              +#+  +:+       +#+        */
+/*   By: cy4gate_mmarinelli <cy4gate_mmarinelli@    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/06 18:13:12 by matteo            #+#    #+#             */
-/*   Updated: 2024/05/16 22:44:29 by matteo           ###   ########.fr       */
+/*   Updated: 2024/06/05 10:18:29 by cy4gate_mma      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ Tile::Tile(): nbr(0)
 	empty = true;
 }
 
-Tile::Tile(uint8_t _nbr): nbr{_nbr}
+Tile::Tile(int _nbr): nbr{_nbr}
 {
 	empty = (0 == nbr);
 }
@@ -52,7 +52,7 @@ bool	Tile::isEmpty() const
 	return empty;
 }
 
-Tile::operator uint8_t() const
+Tile::operator int() const
 {
 	return nbr;
 }
